+++ src/Git_Updater/API/GitHub_API.php
@@ -123,10 +123,13 @@ class GitHub_API extends API implements API_Interface {
 		$endpoint           = '';
 
 		// Release asset.
+		// GitHub will use the release asset URL (browser_download_url) for both updating and installing.
 		if ( $this->use_release_asset( $branch_switch ) ) {
 			$release_asset = $this->get_release_asset();
 
-			return $this->get_release_asset_redirect( $release_asset, true );
+			if ( \property_exists( $this->response['release_asset_response'], 'browser_download_url' ) ) {
+				return $this->response['release_asset_response']->browser_download_url;
+			}
 		}
 
 		/*