+++ src/Git_Updater/Bootstrap.php
@@ -68,9 +68,10 @@ class Bootstrap {
 		\deactivate_plugins( plugin_basename( $this->file ) );
 
 		$message = sprintf(
-			/* translators: %s: documentation URL */
-			__( 'Git Updater is missing required composer dependencies. <a href="%s" target="_blank" rel="noopenernoreferer">Learn more.</a>', 'git-updater' ),
-			'https://github.com/afragen/git-updater/wiki/Installation'
+			/* translators: %1: opening tag, %2: closing tag */
+			__( 'Git Updater is missing required composer dependencies. %1$sLearn more.%2$s', 'git-updater' ),
+			'<a href="https://github.com/afragen/git-updater/wiki/Installation" target="_blank" rel="noreferrer">',
+			'</a>'
 		);
 
 		wp_die( wp_kses_post( $message ) );